v1.0.7 Upgrade (#140)
* feat: add migrations for v1.0.7 upgrade

* feat: add script for simulation of chain upgrade with 1 validator

* feat: add cli command to obtain verification holder by verification id

* fix: fix issue with upgrade handler

* feat: add hack to deploy Arachnid Proxy

* refactor: more meaningful error

* feat: add deployment of Arachnid Deterministic Deployer

* feat: add is_revoked to QueryVerificationsDetailsResponse

* feat: return only non-revoked verification details in x/compliance precompile

* refactor: remove redundant line

* refactor: return in precompile only verifications with active issuers

* refactor: prohibit to convert credential with inactive issuer

* feat: add ffi query to revoke verification

* feat: add precompile handler to revoke verification

* feat: update compliance proxy contract

* test: add test for revocation by issuer

* feat: provide caller address during credential revocation

* feat: add check for caller during credential revocation

* fix: fix issue with invalid non-revocation proof

* feat: add VT_BIOMETRIC verification type

* refactor: remove unused dependencies

* feat: add convertCredential function to precompile interface

* test: add test for converting credential

* feat: add convert credential function signature

* feat: add method to encode convert credential request

* feat: add precompile handler to convert credential

* feat: add cli command to get attached holder public key

* chore: intermediate commit

* feat: add convertCredential method to precompile

* test: fix test for converting v1 verification to v2

* fix: applying fix for pointer problem from https://github.com/CosmWasm/wasmvm/pull/571

* test: add test for non-revocation proof

* refactor: remove unused consts and imports

* refactor: add index variable

* fix: fix build using `make build-docker-local`

* ci: use actions/upload-artifact@v4

* fix: add checks in keeper for max sizes of verification data fields

* test: add test for x/compliance keeper for case with huge original data

* fix: add checks for verification details sizes in compliance_bridge precompile

* feat: start refactoring self attestation to DCAP instead of deprecated EPID

* feat: add DCAP-based self attestation for mainnet / testnet checks

* chore: fix some clippy warnings

* chore: run cargo clippy --fix

+++ sgxvm/src/key_manager/utils.rs
@@ -32,7 +32,7 @@ pub fn random_bytes32() -> SgxResult<[u8; 32]> {
 /// Generates random 32 bytes slice using `sgx_read_rand` function
 pub fn random_nonce() -> SgxResult<[u8; NONCE_SIZE]> {
     let mut buffer = [0u8; NONCE_SIZE];
-    let res = unsafe { sgx_read_rand(&mut buffer as *mut u8, NONCE_SIZE as usize) };
+    let res = unsafe { sgx_read_rand(&mut buffer as *mut u8, NONCE_SIZE) };
 
     if res != sgx_status_t::SGX_SUCCESS {
         println!(
